refactor: split script into multiple files
