+++ fillDelegations/.env
@@ -1,0 +1,2 @@
+FAUCET="0xe5be9a5092b81bca64be81d212e7f2f9eba183bb7a90954f7b76361f6edb5c0a"
+WS_ADDRESS="ws://127.0.0.1:40047"
